+++ src/factory/actionType.ts
@@ -1,25 +1,5 @@
-/**
- * アクションタイプ
- */
-enum ActionType {
-    AuthorizeAction = 'AuthorizeAction',
-    CancelAction = 'CancelAction',
-    CheckAction = 'CheckAction',
-    ConfirmAction = 'ConfirmAction',
-    DeleteAction = 'DeleteAction',
-    GiveAction = 'GiveAction',
-    InformAction = 'InformAction',
-    MoneyTransfer = 'MoneyTransfer',
-    OrderAction = 'OrderAction',
-    PayAction = 'PayAction',
-    PrintAction = 'PrintAction',
-    RefundAction = 'RefundAction',
-    RegisterAction = 'RegisterAction',
-    ReturnAction = 'ReturnAction',
-    SendAction = 'SendAction',
-    UnRegisterAction = 'UnRegisterAction',
-    UpdateAction = 'UpdateAction',
-    UseAction = 'UseAction'
-}
+import { actionType } from '../chevre';
+
+export import ActionType = actionType;
 
 export default ActionType;