+++ src/factory/event.ts
@@ -1,8 +1,6 @@
-import * as OfferFactory from './offer';
-
 import * as chevre from '../chevre';
 
-export type IOffer = OfferFactory.IOffer;
+export type IOffer = chevre.offer.IOffer;
 
 export type IAttributes<T extends chevre.eventType> = chevre.event.IAttributes<T>;