+++ src/factory/offer.ts
@@ -1,6 +1,0 @@
-import * as chevre from '../chevre';
-import * as SeatReservationOfferFactory from './offer/seatReservation';
-
-export import seatReservation = SeatReservationOfferFactory;
-export import IOfferedThrough = chevre.offer.IOfferedThrough;
-export import IOffer = chevre.offer.IOffer;